+++ packages/es-components/src/components/controls/buttons/Button.tsx
@@ -1,13 +1,15 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import styled, { DefaultTheme } from 'styled-components';
+import styled, { DefaultTheme, css } from 'styled-components';
 import tinycolor from 'tinycolor2';
 import { useTheme } from '../../util/useTheme';
 import { darken, getTextColor } from '../../util/colors';
 import ButtonBase, {
   propTypes as buttonBasePropTypes,
   defaultProps as buttonBaseDefaultProps,
-  ButtonBaseProps
+  ButtonBaseProps,
+  UnstyledButton,
+  esComponentsButtonClass
 } from './ButtonBase';
 import type * as CSS from 'csstype';
 import {
@@ -17,6 +19,9 @@ import {
   buttonVariantStyleTypes,
   buttonSizes
 } from 'es-components-shared-types';
+import { PanelButton } from '../../containers/drawer/DrawerPanel';
+import { ScrollIconBaseComponent } from '../../containers/horizontalScrollWrapper/HorizontalScrollWrapper';
+import { BasicProgressButton } from '../../patterns/progress-tracker/progress-tracker-subcomponents';
 
 type BorderRadii = {
   [key in
@@ -26,84 +31,87 @@ type BorderRadii = {
     | 'bottomLeft']: CSS.Property.BorderRadius;
 };
 
-const StyledButton = styled(ButtonBase)<{
+type ButtonStyleProps = {
   colors: ButtonColors;
   borderRadii: BorderRadii;
   mobileBlock: boolean;
   block: boolean;
   buttonSize: ButtonSizeBlock;
-}>`
-  background-color: ${props => props.colors.bgColor};
+};
+
+type ButtonStyleCssProps = {
+  theme: DefaultTheme;
+} & Omit<ButtonStyleProps, 'mobileBlock' | 'block'>;
+
+const getButtonCss = ({
+  theme,
+  colors,
+  borderRadii,
+  buttonSize
+}: ButtonStyleCssProps) => css`
+  background-color: ${colors.bgColor};
   border: 2px solid transparent;
-  border-color: ${props => props.colors.bgColor};
-  border-bottom-left-radius: ${props => props.borderRadii.bottomLeft};
-  border-bottom-right-radius: ${props => props.borderRadii.bottomRight};
-  border-top-left-radius: ${props => props.borderRadii.topLeft};
-  border-top-right-radius: ${props => props.borderRadii.topRight};
+  border-color: ${colors.bgColor};
+  border-bottom-left-radius: ${borderRadii.bottomLeft};
+  border-bottom-right-radius: ${borderRadii.bottomRight};
+  border-top-left-radius: ${borderRadii.topLeft};
+  border-top-right-radius: ${borderRadii.topRight};
   box-sizing: border-box;
-  color: ${props => props.colors.textColor};
+  color: ${colors.textColor};
   cursor: pointer;
-  display: ${props => (props.mobileBlock ? 'block' : 'inline-block')};
   font-family: inherit;
-  font-size: ${props => props.buttonSize.fontSize};
-  font-weight: ${props => props.buttonSize.fontWeight || 'normal'};
-  line-height: ${props =>
-    props.buttonSize.lineHeight
-      ? props.buttonSize.lineHeight
-      : props.theme.font.baseLineHeight};
-  min-width: 100px;
-  padding-bottom: ${props => props.buttonSize.paddingBottom};
-  padding-left: ${props => props.buttonSize.paddingSides};
-  padding-right: ${props => props.buttonSize.paddingSides};
-  padding-top: ${props => props.buttonSize.paddingTop};
+  font-size: ${buttonSize.fontSize};
+  font-weight: ${buttonSize.fontWeight || 'normal'};
+  line-height: ${buttonSize.lineHeight
+    ? buttonSize.lineHeight
+    : theme.font.baseLineHeight};
+  padding-bottom: ${buttonSize.paddingBottom};
+  padding-left: ${buttonSize.paddingSides};
+  padding-right: ${buttonSize.paddingSides};
+  padding-top: ${buttonSize.paddingTop};
   position: relative;
   text-align: center;
   text-decoration: none;
-  text-transform: ${props =>
-    props.buttonSize.textTransform ? props.buttonSize.textTransform : 'none'};
+  text-transform: ${buttonSize.textTransform
+    ? buttonSize.textTransform
+    : 'none'};
   transition: color 0.15s ease-in-out, background-color 0.15s ease-in-out,
     border-color 0.15s ease-in-out, box-shadow 0.15s ease-in-out;
   user-select: none;
   vertical-align: middle;
   white-space: nowrap;
-  width: ${props => (props.mobileBlock ? '100%' : 'auto')};
-
-  @media (min-width: ${props => props.theme.screenSize.tablet}) {
-    display: ${props => (props.block ? 'block' : 'inline-block')};
-    width: ${props => (props.block ? '100%' : 'auto')};
-  }
 
   @media (hover: hover), (-ms-high-contrast: none) {
     &:hover {
-      color: ${props => props.colors.hoverTextColor};
-      background-color: ${props => props.colors.hoverBgColor};
-      border-color: ${props => props.colors.hoverBorderColor};
+      color: ${colors.hoverTextColor};
+      background-color: ${colors.hoverBgColor};
+      border-color: ${colors.hoverBorderColor};
       text-decoration: none;
     }
   }
 
   &:focus {
-    color: ${props => props.colors.hoverTextColor};
-    background-color: ${props => props.colors.hoverBgColor};
+    color: ${colors.hoverTextColor};
+    background-color: ${colors.hoverBgColor};
     border: 2px solid;
-    border-color: ${props => props.colors.hoverBorderColor};
+    border-color: ${colors.hoverBorderColor};
     box-shadow: 0 1px 1px rgba(0, 0, 0, 0.075),
-      0 0 0 0.2rem ${props => props.colors.focusBoxShadowColor};
+      0 0 0 0.2rem ${colors.focusBoxShadowColor};
     outline: none;
   }
 
   &:active,
   &.pressed {
-    color: ${props => props.colors.activeTextColor};
-    background-color: ${props => props.colors.activeBgColor};
-    border-color: ${props => props.colors.activeBorderColor};
+    color: ${colors.activeTextColor};
+    background-color: ${colors.activeBgColor};
+    border-color: ${colors.activeBorderColor};
     box-shadow: inset 0 3px 5px rgba(0, 0, 0, 0.25);
   }
 
   &:active:focus,
   &.pressed:focus {
     box-shadow: inset 0 3px 5px rgba(0, 0, 0, 0.25),
-      0 0 0 0.2rem ${props => props.colors.focusBoxShadowColor};
+      0 0 0 0.2rem ${colors.focusBoxShadowColor};
   }
 
   &[disabled],
@@ -120,7 +128,7 @@ const StyledButton = styled(ButtonBase)<{
       color: ${({ theme }) => theme.colors.gray6};
 
       &:focus {
-        box-shadow: 0 0 0 0.2rem ${props => props.colors.focusBoxShadowColor};
+        box-shadow: 0 0 0 0.2rem ${colors.focusBoxShadowColor};
       }
     }
 
@@ -129,6 +137,79 @@ const StyledButton = styled(ButtonBase)<{
     }
   }
 `;
+
+const StyledButton = styled(ButtonBase)<ButtonStyleProps>`
+  ${getButtonCss}
+  min-width: 100px;
+  display: ${({ mobileBlock }) => (mobileBlock ? 'block' : 'inline-block')};
+  width: ${({ mobileBlock }) => (mobileBlock ? '100%' : 'auto')};
+
+  @media (min-width: ${({ theme }) => theme.screenSize.tablet}) {
+    display: ${({ block }) => (block ? 'block' : 'inline-block')};
+    width: ${({ block }) => (block ? '100%' : 'auto')};
+  }
+`;
+
+const getBorderRadii = (buttonSize: ButtonSizeBlock): BorderRadii => ({
+  topLeft: buttonSize.borderRadius,
+  topRight: buttonSize.borderRadius,
+  bottomRight: buttonSize.borderRadius,
+  bottomLeft: buttonSize.borderRadius
+});
+
+const onlyHasOldButtons = () =>
+  [...document.querySelectorAll('button')].every(
+    button => !button.classList.contains(esComponentsButtonClass)
+  );
+
+export const globalButtonCss = css`
+  ${({ theme }) =>
+    !onlyHasOldButtons() &&
+    css`
+      form
+        button:not(
+          .react-datepicker__navigation,
+          [name^='rsg'],
+          [class^='rsg'],
+          ${PanelButton},
+            ${ScrollIconBaseComponent},
+            ${StyledButton},
+            ${BasicProgressButton},
+            ${UnstyledButton},
+            .${esComponentsButtonClass}
+        ) {
+        ${getButtonCss({
+          theme,
+          colors: getButtonColors(theme, false, 'default'),
+          borderRadii: getBorderRadii(theme.buttonStyles.button.size.default),
+          buttonSize: theme.buttonStyles.button.size.default
+        })}
+        ${buttonVariantStyleTypes.map(style =>
+          buttonSizes.map(size => {
+            const buttonSize = theme.buttonStyles.button.size[size];
+            const buttonColors = getButtonColors(theme, false, style);
+            const borderRadii = getBorderRadii(buttonSize);
+            return css`
+              &.${[style, size].join('.')} {
+                ${getButtonCss({
+                  theme,
+                  colors: buttonColors,
+                  borderRadii,
+                  buttonSize
+                })}
+              }
+            `;
+          })
+        )}
+      }
+    `}
+`;
+
+export const buttonStyleTypes = [
+  ...buttonVariantStyleTypes,
+  'inherited'
+] as const;
+export type ButtonStyleType = (typeof buttonStyleTypes)[number];
 
 const buttonColorProps = [
   'bgColor',
@@ -152,10 +233,19 @@ type GuaranteedButtonVariant = Omit<ButtonVariant, 'bgColor'> & {
   bgColor: NonNullable<CSS.Property.BackgroundColor>;
 };
 
+function getButtonVariant(
+  theme: DefaultTheme,
+  buttonVariant: ButtonStyleType
+): GuaranteedButtonVariant {
+  const variant = theme.buttonStyles.button.variant[buttonVariant] || {};
+  variant.bgColor = variant.bgColor || theme.colors.gray4;
+  return variant;
+}
+
 function getButtonColors<T extends boolean>(
   theme: DefaultTheme,
   isInheritedStyle: T,
-  buttonVariant?: T extends false ? GuaranteedButtonVariant : undefined
+  buttonVariant: ButtonStyleType = 'default'
 ): ButtonColors {
   const baseButtonColors: ButtonColors = {
     bgColor: 'inherited',
@@ -174,7 +264,7 @@ function getButtonColors<T extends boolean>(
     return baseButtonColors;
   }
 
-  const variant = buttonVariant as GuaranteedButtonVariant;
+  const variant = getButtonVariant(theme, buttonVariant);
 
   const focusBoxShadowColor = tinycolor
     .mix(variant.bgColor, theme.colors.black, 14)
@@ -208,11 +298,6 @@ function getButtonColors<T extends boolean>(
   return calculatedButtonColors;
 }
 
-export const buttonStyleTypes = [
-  ...buttonVariantStyleTypes,
-  'inherited'
-] as const;
-export type ButtonStyleType = (typeof buttonStyleTypes)[number];
 export type ButtonProps = ButtonBaseProps & {
   size?: ButtonSize;
   styleType?: ButtonStyleType;
@@ -237,9 +322,6 @@ const Button = React.forwardRef<HTMLButtonElement, ButtonProps>(function Button(
 ) {
   const theme = useTheme();
   const buttonSize = theme.buttonStyles.button.size[size];
-  const variant = theme.buttonStyles.button.variant[
-    styleType
-  ] as GuaranteedButtonVariant;
   const isInheritedStyle = styleType === 'inherited';
   const mobileBlockSetting =
     flatLeftEdge || flatRightEdge ? false : mobileBlock;
@@ -252,7 +334,7 @@ const Button = React.forwardRef<HTMLButtonElement, ButtonProps>(function Button(
     bottomLeft: flatLeftEdge ? 0 : defaultRadius
   };
 
-  const buttonColors = getButtonColors(theme, isInheritedStyle, variant);
+  const buttonColors = getButtonColors(theme, isInheritedStyle, styleType);
 
   return (
     <StyledButton