+++ packages/es-components/src/components/controls/buttons/ButtonBase.tsx
@@ -9,7 +9,7 @@ import { useMonitoringCallback } from '../../../hooks/useMonitoringHooks';
 
 // Using this because React does not like many of our upstream props.
 // This allows that to flow without issue.
-const UnstyledButton = styled.button`
+export const UnstyledButton = styled.button`
   // noop
 `;
 
@@ -20,18 +20,27 @@ export type ButtonBaseProps = Override<
   }
 >;
 
+export const esComponentsButtonClass = 'es-components-button';
 const ButtonBase = React.forwardRef<HTMLButtonElement, ButtonBaseProps>(
-  function ButtonBaseInner({ waiting, onClick: onClickProp, ...props }, ref) {
+  function ButtonBaseInner(
+    { waiting, onClick: onClickProp, className, ...props },
+    ref
+  ) {
     // eslint-disable-next-line @typescript-eslint/no-unused-vars
-    const onClick: React.MouseEventHandler<HTMLButtonElement> =
-      useMonitoringCallback(
-        (currentOnClick, ...args) => !waiting && currentOnClick?.(...args),
-        [waiting],
-        onClickProp
-      );
+    const onClick = useMonitoringCallback(
+      (
+        currentOnClick,
+        ...args: Parameters<React.MouseEventHandler<HTMLButtonElement>>
+      ) => !waiting && currentOnClick?.(...args),
+      [waiting],
+      onClickProp
+    );
 
     const computedProps = {
       ...props,
+      ...(className
+        ? { className: `${className} ${esComponentsButtonClass}` }
+        : { className: esComponentsButtonClass }),
       ...(waiting ? { 'data-waiting': waiting } : {})
     };