+++ packages/es-components/src/components/navigation/Anchor.tsx
@@ -1,6 +1,6 @@
-import styled from 'styled-components';
+import styled, { css } from 'styled-components';
 
-const Anchor = styled.a`
+const anchorCss = css`
   background-color: transparent;
   box-sizing: border-box;
   color: ${props => props.theme.colors.primary};
@@ -10,6 +10,16 @@ const Anchor = styled.a`
     color: ${props => props.theme.colors.primaryHover};
     text-decoration: none;
   }
+`;
+
+export const globalAnchorCss = css`
+  a {
+    ${anchorCss};
+  }
+`;
+
+const Anchor = styled.a`
+  ${anchorCss}
 `;
 
 /** @component */