+++ packages/es-components/src/components/patterns/progress-tracker/progress-tracker-subcomponents.tsx
@@ -70,7 +70,7 @@ export const ProgressContainer = styled.ol<ProgressContainerProps>`
   }
 `;
 
-const BasicProgressButton = styled.button`
+export const BasicProgressButton = styled.button`
   align-items: center;
   background: none;
   border: none;