+++ packages/es-components/src/components/util/style-utils.tsx
@@ -1,0 +1,9 @@
+import { css } from 'styled-components';
+
+export const baseFontCss = css`
+  ${({ theme }) =>
+    theme?.font &&
+    css`
+      font-family: ${theme.font.baseFontFace};
+    `}
+`;